src/hybrid/mves/rabbitmq/pyproject.toml
[project]
name = "rabbitmq-docker-python"
version = "0.1.0"
description = "Minimal viable example for RabbitMQ with Docker and Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pika>=1.3.2",
    "python-dotenv>=1.0.0",
]